requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-jwt
gunicorn
ipython
mysql-connector
mysqlclient
pandas
pydruid
python-dotenv